feat: fetch部実装中

+++ src/api/repositories/card.repository.ts
@@ -1,6 +1,8 @@
 import { Card } from '../models/card.model'
 import { apiClient } from '../clients/apiClient'
 import { Card as CardResponse } from '../schemas/generated/schemas'
+import { getServerSession } from 'next-auth'
+import { authOptions } from '@/pages/api/auth/[...nextauth]'
 
 export interface CardRepository {
   listCards: () => Promise<CardResponse[]>
@@ -13,14 +15,22 @@ export interface CardRepository {
 }
 
 const listCards: CardRepository['listCards'] = async (): Promise<CardResponse[]> => {
-  const data = await apiClient.get('/cards')
+  const session = await getServerSession(authOptions)
+  if (!session) {
+    throw new Error('Session is not found')
+  }
+  const data = await apiClient.get('/cards',{accessToken: session.user.access_token || ""},)
   return data
 }
 
 const createCard: CardRepository['createCard'] = async (
   cardData: Pick<Card, 'title' | 'content' | 'period'>,
 ): Promise<CardResponse> => {
-  const data = await apiClient.post('/cards', {
+  const session = await getServerSession(authOptions)
+  if (!session) {
+    throw new Error('Session is not found')
+  }
+  const data = await apiClient.post('/cards',{accessToken: session.user.access_token || ""}, {
     period: cardData.period,
     title: cardData.title,
     content: cardData.content,
@@ -31,7 +41,11 @@ const createCard: CardRepository['createCard'] = async (
 const createCards: CardRepository['createCards'] = async (
   cardData: Pick<Card, 'title' | 'content' | 'period'>[],
 ): Promise<CardResponse[]> => {
-  const data = await apiClient.post(`/cards/batch`, {
+  const session = await getServerSession(authOptions)
+  if (!session) {
+    throw new Error('Session is not found')
+  }
+  const data = await apiClient.post(`/cards/batch`,{accessToken: session.user.access_token || ""}, {
     cards: cardData,
   })
   return data
@@ -40,7 +54,11 @@ const createCards: CardRepository['createCards'] = async (
 const updateCard: CardRepository['updateCard'] = async (
   cardData: Partial<Card>,
 ): Promise<CardResponse> => {
-  const data = await apiClient.patch(`/cards/${cardData.id}`, {
+  const session = await getServerSession(authOptions)
+  if (!session) {
+    throw new Error('Session is not found')
+  }
+  const data = await apiClient.patch(`/cards/${cardData.id}`,{accessToken: session.user.access_token || ""},  {
     // cardDataに保持している値を展開してしまう。
     ...cardData,
   })
@@ -48,7 +66,11 @@ const updateCard: CardRepository['updateCard'] = async (
 }
 
 const deleteCard: CardRepository['deleteCard'] = async (cardData: Pick<Card, 'id'>) => {
-  await apiClient.destroy(`/cards/${cardData.id}`)
+  const session = await getServerSession(authOptions)
+  if (!session) {
+    throw new Error('Session is not found')
+  }
+  await apiClient.destroy(`/cards/${cardData.id}`,{accessToken: session.user.access_token || ""}, )
 }
 
 export const cardRepository: CardRepository = {
